+++ profile-widget/profilescene.cpp
@@ -1,8 +1,19 @@
 // SPDX-License-Identifier: GPL-2.0
 #include "profilescene.h"
+#include "diveeventitem.h"
 #include "divecartesianaxis.h"
+#include "diveprofileitem.h"
+#include "divetextitem.h"
+#include "tankitem.h"
+#include "core/device.h"
+#include "core/event.h"
 #include "core/pref.h"
+#include "core/profile.h"
+#include "core/qthelper.h"	// for decoMode()
+#include "core/subsurface-string.h"
+#include "core/settings/qPrefDisplay.h"
 #include "qt-models/diveplotdatamodel.h"
+#include "qt-models/diveplannermodel.h"
 
 const static struct ProfileItemPos {
 	struct Pos {
@@ -14,6 +25,8 @@ const static struct ProfileItemPos {
 		QLineF expanded;
 		QLineF intermediate;
 	};
+	Pos dcLabel;
+	Pos tankBar;
 	Axis depth;
 	Axis partialPressure;
 	Axis partialPressureTissue;
@@ -29,10 +42,34 @@ const static struct ProfileItemPos {
 	ProfileItemPos();
 } itemPos;
 
+template<typename T, class... Args>
+T *ProfileScene::createItem(const DiveCartesianAxis &vAxis, int vColumn, int z, Args&&... args)
+{
+	T *res = new T(*dataModel, *timeAxis, DivePlotDataModel::TIME, vAxis, vColumn,
+		       std::forward<Args>(args)...);
+	res->setZValue(static_cast<double>(z));
+	profileItems.push_back(res);
+	return res;
+}
+
+PartialPressureGasItem *ProfileScene::createPPGas(int column, color_index_t color, color_index_t colorAlert,
+						    const double *thresholdSettingsMin, const double *thresholdSettingsMax)
+{
+	PartialPressureGasItem *item = createItem<PartialPressureGasItem>(*gasYAxis, column, 99, fontPrintScale);
+	item->setThresholdSettingsKey(thresholdSettingsMin, thresholdSettingsMax);
+	item->setColors(getColor(color, isGrayscale), getColor(colorAlert, isGrayscale));
+	return item;
+}
+
 ProfileScene::ProfileScene(double fontPrintScale) :
+	animSpeed(0),
+	d(nullptr),
+	dc(-1),
 	fontPrintScale(fontPrintScale),
 	printMode(false),
 	isGrayscale(false),
+	maxtime(-1),
+	maxdepth(-1),
 	dataModel(new DivePlotDataModel(this)),
 	profileYAxis(new DepthAxis(fontPrintScale, *this)),
 	gasYAxis(new PartialGasPressureAxis(*dataModel, fontPrintScale, *this)),
@@ -40,8 +77,28 @@ ProfileScene::ProfileScene(double fontPrintScale) :
 	timeAxis(new TimeAxis(fontPrintScale, *this)),
 	cylinderPressureAxis(new DiveCartesianAxis(fontPrintScale, *this)),
 	heartBeatAxis(new DiveCartesianAxis(fontPrintScale, *this)),
-	percentageAxis(new DiveCartesianAxis(fontPrintScale, *this))
+	percentageAxis(new DiveCartesianAxis(fontPrintScale, *this)),
+	diveProfileItem(createItem<DiveProfileItem>(*profileYAxis, DivePlotDataModel::DEPTH, 0, fontPrintScale)),
+	temperatureItem(createItem<DiveTemperatureItem>(*temperatureAxis, DivePlotDataModel::TEMPERATURE, 1, fontPrintScale)),
+	meanDepthItem(createItem<DiveMeanDepthItem>(*profileYAxis, DivePlotDataModel::INSTANT_MEANDEPTH, 1, fontPrintScale)),
+	gasPressureItem(createItem<DiveGasPressureItem>(*cylinderPressureAxis, DivePlotDataModel::TEMPERATURE, 1, fontPrintScale)),
+	diveComputerText(new DiveTextItem(fontPrintScale)),
+	reportedCeiling(createItem<DiveReportedCeiling>(*profileYAxis, DivePlotDataModel::CEILING, 1, fontPrintScale)),
+	pn2GasItem(createPPGas(DivePlotDataModel::PN2, PN2, PN2_ALERT, NULL, &prefs.pp_graphs.pn2_threshold)),
+	pheGasItem(createPPGas(DivePlotDataModel::PHE, PHE, PHE_ALERT, NULL, &prefs.pp_graphs.phe_threshold)),
+	po2GasItem(createPPGas(DivePlotDataModel::PO2, PO2, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	o2SetpointGasItem(createPPGas(DivePlotDataModel::O2SETPOINT, O2SETPOINT, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	ccrsensor1GasItem(createPPGas(DivePlotDataModel::CCRSENSOR1, CCRSENSOR1, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	ccrsensor2GasItem(createPPGas(DivePlotDataModel::CCRSENSOR2, CCRSENSOR2, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	ccrsensor3GasItem(createPPGas(DivePlotDataModel::CCRSENSOR3, CCRSENSOR3, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	ocpo2GasItem(createPPGas(DivePlotDataModel::SCR_OC_PO2, SCR_OCPO2, PO2_ALERT, &prefs.pp_graphs.po2_threshold_min, &prefs.pp_graphs.po2_threshold_max)),
+	diveCeiling(createItem<DiveCalculatedCeiling>(*profileYAxis, DivePlotDataModel::CEILING, 1, fontPrintScale)),
+	decoModelParameters(new DiveTextItem(fontPrintScale)),
+	heartBeatItem(createItem<DiveHeartrateItem>(*heartBeatAxis, DivePlotDataModel::HEARTBEAT, 1, fontPrintScale)),
+	tankItem(new TankItem(*timeAxis, fontPrintScale))
 {
+	init_plot_info(&plotInfo);
+
 	setSceneRect(0, 0, 100, 100);
 	setItemIndexMethod(QGraphicsScene::NoIndex);
 
@@ -96,8 +153,31 @@ ProfileScene::ProfileScene(double fontPrintScale) :
 	timeAxis->setLinesVisible(true);
 	profileYAxis->setLinesVisible(true);
 	gasYAxis->setZValue(timeAxis->zValue() + 1);
+	tankItem->setZValue(100);
+
+	// show the deco model parameters at the top in the center
+	decoModelParameters->setY(0);
+	decoModelParameters->setX(50);
+	decoModelParameters->setBrush(getColor(PRESSURE_TEXT));
+	decoModelParameters->setAlignment(Qt::AlignHCenter | Qt::AlignBottom);
+
+	diveComputerText->setAlignment(Qt::AlignRight | Qt::AlignTop);
+	diveComputerText->setBrush(getColor(TIME_TEXT, isGrayscale));
+	diveComputerText->setPos(itemPos.dcLabel.on);
+
+	tankItem->setPos(itemPos.tankBar.on);
+
+	for (int i = 0; i < 16; i++) {
+		DiveCalculatedTissue *tissueItem = createItem<DiveCalculatedTissue>(*profileYAxis, DivePlotDataModel::TISSUE_1 + i, i + 1, fontPrintScale);
+		allTissues.append(tissueItem);
+		DivePercentageItem *percentageItem = createItem<DivePercentageItem>(*percentageAxis, DivePlotDataModel::PERCENTAGE_1 + i, i + 1, i, fontPrintScale);
+		allPercentages.append(percentageItem);
+	}
 
 	// Add items to scene
+	addItem(diveComputerText);
+	addItem(tankItem);
+	addItem(decoModelParameters);
 	addItem(profileYAxis);
 	addItem(gasYAxis);
 	addItem(temperatureAxis);
@@ -105,15 +185,65 @@ ProfileScene::ProfileScene(double fontPrintScale) :
 	addItem(cylinderPressureAxis);
 	addItem(percentageAxis);
 	addItem(heartBeatAxis);
+
+	for (AbstractProfilePolygonItem *item: profileItems)
+		addItem(item);
 }
 
 ProfileScene::~ProfileScene()
 {
+	free_plot_info_data(&plotInfo);
+}
+
+void ProfileScene::clear()
+{
+	dataModel->clear();
+
+	for (AbstractProfilePolygonItem *item: profileItems)
+		item->clear();
+
+	// the events will have connected slots which can fire after
+	// the dive and its data have been deleted - so explictly delete
+	// the DiveEventItems
+	qDeleteAll(eventItems);
+	eventItems.clear();
 }
 
 static bool ppGraphsEnabled()
 {
 	return prefs.pp_graphs.po2 || prefs.pp_graphs.pn2 || prefs.pp_graphs.phe;
+}
+
+// Update visibility of non-interactive chart features according to preferences
+void ProfileScene::updateVisibility()
+{
+#ifndef SUBSURFACE_MOBILE
+	pn2GasItem->setVisible(prefs.pp_graphs.pn2);
+	po2GasItem->setVisible(prefs.pp_graphs.po2);
+	pheGasItem->setVisible(prefs.pp_graphs.phe);
+
+	const struct divecomputer *currentdc = d ? get_dive_dc_const(d, dc) : nullptr;
+	bool setpointflag = currentdc && currentdc->divemode == CCR && prefs.pp_graphs.po2;
+	bool sensorflag = setpointflag && prefs.show_ccr_sensors;
+	o2SetpointGasItem->setVisible(setpointflag && prefs.show_ccr_setpoint);
+	ccrsensor1GasItem->setVisible(sensorflag);
+	ccrsensor2GasItem->setVisible(currentdc && sensorflag && currentdc->no_o2sensors > 1);
+	ccrsensor3GasItem->setVisible(currentdc && sensorflag && currentdc->no_o2sensors > 2);
+	ocpo2GasItem->setVisible(currentdc && currentdc->divemode == PSCR && prefs.show_scr_ocpo2);
+
+	heartBeatItem->setVisible(prefs.hrgraph);
+#endif
+	diveCeiling->setVisible(prefs.calcceiling);
+	decoModelParameters->setVisible(prefs.calcceiling);
+#ifndef SUBSURFACE_MOBILE
+	for (DiveCalculatedTissue *tissue: allTissues)
+		tissue->setVisible(prefs.calcalltissues && prefs.calcceiling);
+	for (DivePercentageItem *percentage: allPercentages)
+		percentage->setVisible(prefs.percentagegraph);
+#endif
+	meanDepthItem->setVisible(prefs.show_average_depth);
+	reportedCeiling->setVisible(prefs.dcceiling);
+	tankItem->setVisible(prefs.tankbar);
 }
 
 void ProfileScene::updateAxes()
@@ -196,6 +326,18 @@ ProfileItemPos::ProfileItemPos()
 	// Scene is *always* (double) 100 / 100.
 	// TODO: The relative scaling doesn't work well with large or small
 	// profiles and needs to be fixed.
+
+	dcLabel.on.setX(3);
+	dcLabel.on.setY(100);
+	dcLabel.off.setX(-10);
+	dcLabel.off.setY(100);
+
+	tankBar.on.setX(0);
+#ifndef SUBSURFACE_MOBILE
+	tankBar.on.setY(91.95);
+#else
+	tankBar.on.setY(86.4);
+#endif
 
 	//Depth Axis Config
 	depth.on.setX(3);
@@ -294,4 +436,229 @@ ProfileItemPos::ProfileItemPos()
 	percentage.expanded.setP2(QPointF(0, 15));
 	percentageWithTankBar = percentage;
 	percentageWithTankBar.expanded.setP2(QPointF(0, 11.9));
+}
+
+void ProfileScene::plotDive(const struct dive *dIn, int dcIn, DivePlannerPointsModel *plannerModel, bool inPlanner, bool instant, bool calcMax)
+{
+	d = dIn;
+	dc = dcIn;
+	if (!d) {
+		clear();
+		return;
+	}
+
+	if (!plannerModel) {
+		if (decoMode(false) == VPMB)
+			decoModelParameters->setText(QString("VPM-B +%1").arg(prefs.vpmb_conservatism));
+		else
+			decoModelParameters->setText(QString("GF %1/%2").arg(prefs.gflow).arg(prefs.gfhigh));
+	} else {
+		struct diveplan &diveplan = plannerModel->getDiveplan();
+		if (decoMode(inPlanner) == VPMB)
+			decoModelParameters->setText(QString("VPM-B +%1").arg(diveplan.vpmb_conservatism));
+		else
+			decoModelParameters->setText(QString("GF %1/%2").arg(diveplan.gflow).arg(diveplan.gfhigh));
+	}
+
+	const struct divecomputer *currentdc = get_dive_dc_const(d, dc);
+	if (!currentdc || !currentdc->samples)
+		return;
+
+	animSpeed = instant || printMode ? 0 : qPrefDisplay::animation_speed();
+
+	bool setpointflag = (currentdc->divemode == CCR) && prefs.pp_graphs.po2;
+	bool sensorflag = setpointflag && prefs.show_ccr_sensors;
+	o2SetpointGasItem->setVisible(setpointflag && prefs.show_ccr_setpoint);
+	ccrsensor1GasItem->setVisible(sensorflag);
+	ccrsensor2GasItem->setVisible(sensorflag && (currentdc->no_o2sensors > 1));
+	ccrsensor3GasItem->setVisible(sensorflag && (currentdc->no_o2sensors > 2));
+	ocpo2GasItem->setVisible((currentdc->divemode == PSCR) && prefs.show_scr_ocpo2);
+
+	updateVisibility();
+
+	// A non-null planner_ds signals to create_plot_info_new that the dive is currently planned.
+	struct deco_state *planner_ds = inPlanner && plannerModel ? &plannerModel->final_deco_state : nullptr;
+
+	/* This struct holds all the data that's about to be plotted.
+	 * I'm not sure this is the best approach ( but since we are
+	 * interpolating some points of the Dive, maybe it is... )
+	 * The  Calculation of the points should be done per graph,
+	 * so I'll *not* calculate everything if something is not being
+	 * shown.
+	 * create_plot_info_new() automatically frees old plot data.
+	 */
+	create_plot_info_new(d, get_dive_dc_const(d, dc), &plotInfo, !calcMax, planner_ds);
+
+	int newMaxtime = get_maxtime(&plotInfo);
+	if (calcMax || newMaxtime > maxtime)
+		maxtime = newMaxtime;
+
+	/* Only update the max. depth if it's bigger than the current ones
+	 * when we are dragging the handler to plan / add dive.
+	 * otherwhise, update normally.
+	 */
+	int newMaxDepth = get_maxdepth(&plotInfo);
+	if (!calcMax) {
+		if (maxdepth < newMaxDepth) {
+			maxdepth = newMaxDepth;
+		}
+	} else {
+		maxdepth = newMaxDepth;
+	}
+
+	dataModel->setDive(plotInfo);
+
+	// It seems that I'll have a lot of boilerplate setting the model / axis for
+	// each item, I'll mostly like to fix this in the future, but I'll keep at this for now.
+	profileYAxis->setMaximum(maxdepth);
+	profileYAxis->updateTicks();
+
+	temperatureAxis->setMinimum(plotInfo.mintemp);
+	temperatureAxis->setMaximum(plotInfo.maxtemp - plotInfo.mintemp > 2000 ? plotInfo.maxtemp : plotInfo.mintemp + 2000);
+
+	if (plotInfo.maxhr) {
+		int heartBeatAxisMin = lrint(plotInfo.minhr / 5.0 - 0.5) * 5;
+		int heartBeatAxisMax, heartBeatAxisTick;
+		if (plotInfo.maxhr - plotInfo.minhr < 40)
+			heartBeatAxisTick = 10;
+		else if (plotInfo.maxhr - plotInfo.minhr < 80)
+			heartBeatAxisTick = 20;
+		else if (plotInfo.maxhr - plotInfo.minhr < 100)
+			heartBeatAxisTick = 25;
+		else
+			heartBeatAxisTick = 50;
+		for (heartBeatAxisMax = heartBeatAxisMin; heartBeatAxisMax < plotInfo.maxhr; heartBeatAxisMax += heartBeatAxisTick);
+		heartBeatAxis->setMinimum(heartBeatAxisMin);
+		heartBeatAxis->setMaximum(heartBeatAxisMax + 1);
+		heartBeatAxis->setTickInterval(heartBeatAxisTick);
+		heartBeatAxis->updateTicks(HR_AXIS); // this shows the ticks
+	}
+	heartBeatAxis->setVisible(prefs.hrgraph && plotInfo.maxhr);
+
+	percentageAxis->setMinimum(0);
+	percentageAxis->setMaximum(100);
+	percentageAxis->setVisible(false);
+	percentageAxis->updateTicks(HR_AXIS);
+
+	if (calcMax)
+		timeAxis->setMaximum(maxtime);
+
+	int i, incr;
+	static int increments[8] = { 10, 20, 30, 60, 5 * 60, 10 * 60, 15 * 60, 30 * 60 };
+	/* Time markers: at most every 10 seconds, but no more than 12 markers.
+	 * We start out with 10 seconds and increment up to 30 minutes,
+	 * depending on the dive time.
+	 * This allows for 6h dives - enough (I hope) for even the craziest
+	 * divers - but just in case, for those 8h depth-record-breaking dives,
+	 * we double the interval if this still doesn't get us to 12 or fewer
+	 * time markers */
+	i = 0;
+	while (i < 7 && maxtime / increments[i] > 12)
+		i++;
+	incr = increments[i];
+	while (maxtime / incr > 12)
+		incr *= 2;
+	timeAxis->setTickInterval(incr);
+	timeAxis->updateTicks();
+	cylinderPressureAxis->setMinimum(plotInfo.minpressure);
+	cylinderPressureAxis->setMaximum(plotInfo.maxpressure);
+
+#ifdef SUBSURFACE_MOBILE
+	if (currentdc->divemode == CCR) {
+		gasYAxis->setPos(itemPos.partialPressure.on);
+		gasYAxis->setLine(itemPos.partialPressure.expanded);
+
+		tankItem->setVisible(false);
+		pn2GasItem->setVisible(false);
+		po2GasItem->setVisible(prefs.pp_graphs.po2);
+		pheGasItem->setVisible(false);
+		o2SetpointGasItem->setVisible(prefs.show_ccr_setpoint);
+		ccrsensor1GasItem->setVisible(prefs.show_ccr_sensors);
+		ccrsensor2GasItem->setVisible(prefs.show_ccr_sensors && (currentdc->no_o2sensors > 1));
+		ccrsensor3GasItem->setVisible(prefs.show_ccr_sensors && (currentdc->no_o2sensors > 1));
+		ocpo2GasItem->setVisible((currentdc->divemode == PSCR) && prefs.show_scr_ocpo2);
+		//when no gas graph, we can show temperature
+		if (!po2GasItem->isVisible() &&
+		    !o2SetpointGasItem->isVisible() &&
+		    !ccrsensor1GasItem->isVisible() &&
+		    !ccrsensor2GasItem->isVisible() &&
+		    !ccrsensor3GasItem->isVisible() &&
+		    !ocpo2GasItem->isVisible())
+			temperatureItem->setVisible(true);
+		else
+			temperatureItem->setVisible(false);
+	} else {
+		tankItem->setVisible(prefs.tankbar);
+		gasYAxis->setPos(itemPos.partialPressure.off);
+		pn2GasItem->setVisible(false);
+		po2GasItem->setVisible(false);
+		pheGasItem->setVisible(false);
+		o2SetpointGasItem->setVisible(false);
+		ccrsensor1GasItem->setVisible(false);
+		ccrsensor2GasItem->setVisible(false);
+		ccrsensor3GasItem->setVisible(false);
+		ocpo2GasItem->setVisible(false);
+	}
+#endif
+	tankItem->setData(&plotInfo, d);
+
+	gasYAxis->update();
+
+	// Replot dive items
+	for (AbstractProfilePolygonItem *item: profileItems)
+		item->replot(d, inPlanner);
+
+	// The event items are a bit special since we don't know how many events are going to
+	// exist on a dive, so I cant create cache items for that. that's why they are here
+	// while all other items are up there on the constructor.
+	qDeleteAll(eventItems);
+	eventItems.clear();
+	struct event *event = currentdc->events;
+	struct gasmix lastgasmix = get_gasmix_at_time(d, get_dive_dc_const(d, dc), duration_t{1});
+
+	while (event) {
+#ifndef SUBSURFACE_MOBILE
+		// if print mode is selected only draw headings, SP change, gas events or bookmark event
+		if (printMode) {
+			if (empty_string(event->name) ||
+			    !(strcmp(event->name, "heading") == 0 ||
+			      (same_string(event->name, "SP change") && event->time.seconds == 0) ||
+			      event_is_gaschange(event) ||
+			      event->type == SAMPLE_EVENT_BOOKMARK)) {
+				event = event->next;
+				continue;
+			}
+		}
+#else
+		// printMode is always selected for SUBSURFACE_MOBILE due to font problems
+		// BUT events are wanted.
+#endif
+		DiveEventItem *item = new DiveEventItem(d, event, lastgasmix, dataModel,
+							timeAxis, profileYAxis, animSpeed,
+							fontPrintScale);
+		item->setZValue(2);
+		addItem(item);
+		eventItems.push_back(item);
+		if (event_is_gaschange(event))
+			lastgasmix = get_gasmix_from_event(d, event);
+		event = event->next;
+	}
+
+	// Only set visible the events that should be visible
+	Q_FOREACH (DiveEventItem *event, eventItems) {
+		event->setVisible(!event->shouldBeHidden());
+	}
+	QString dcText = get_dc_nickname(currentdc);
+	if (dcText == "planned dive")
+		dcText = tr("Planned dive");
+	else if (dcText == "manually added dive")
+		dcText = tr("Manually added dive");
+	else if (dcText.isEmpty())
+		dcText = tr("Unknown dive computer");
+#ifndef SUBSURFACE_MOBILE
+	int nr;
+	if ((nr = number_of_computers(d)) > 1)
+		dcText += tr(" (#%1 of %2)").arg(dc + 1).arg(nr);
+#endif
+	diveComputerText->setText(dcText);
 }
